Fix for mondo bungling the 'exactMatch' property
MONDO had historically included 'exactMatch' as a distinct property,
viewable in the OLS JSON. The data previously included here now shows
up as a list of strings, with keys separated by spaces, under the
property_value key. Added code to read this key and add the contained
strings back to the annotation key as it was before.

I expect this is an error in the data produced by MONDO and that they
will fix it eventually. This change should not break after MONDO gets
their act together, but it is reasonable to remove it after their data
reverts to its previous form.